パーサーのヘッダー分割
