+++ WorkFlow/ASTree/Parser/Parser/Operators.hpp
@@ -1,0 +1,19 @@
+//
+//  Operators.hpp
+//  WorkFlow
+//
+//  Created by yuki on 2019/07/31.
+//  Copyright © 2019 yuki. All rights reserved.
+//
+
+#ifndef Operators_hpp
+#define Operators_hpp
+
+#include <set>
+#include <string>
+
+class Operators {
+    std::set<std::string>
+}
+
+#endif /* Operators_hpp */
