+++ WorkFlow/ASTree/Parser/_AnyParser.hpp
@@ -1,0 +1,13 @@
+//
+//  _AnyParser.hpp
+//  WorkFlow
+//
+//  Created by yuki on 2019/07/31.
+//  Copyright © 2019 yuki. All rights reserved.
+//
+
+#ifndef _AnyParser_h
+#define _AnyParser_h
+
+
+#endif /* _AnyParser_h */
